feat: additional MyPy error reduction - 15 more errors fixed (341→326)
## Progress Update
This commit continues the systematic MyPy error reduction, fixing an additional
15 errors and bringing the total fixed to **229 errors** (from 555 to 326).

## Error Categories Fixed This Session

### Assignment Errors (8 fixed)
- Fixed Pydantic Field default_factory usage with lambda constructors
- Added explicit type annotations for variable assignments
- Fixed Union type variable declarations in neural networks
- Improved Optional Path handling in model storage

### Union-attr Errors (4 fixed)
- Added None checks for Future objects before method calls
- Fixed conditional attribute access patterns
- Improved type safety for optional objects

### Call-arg Errors (2 fixed)
- Fixed Pydantic model constructor arguments
- Added explicit error parameters to ApiResponse calls

### Index Errors (1 fixed)
- Added type ignore for dynamic dictionary indexing in visualization

## Key Files Updated
- **ktrdr/cli/progress.py**: Fixed TaskID type annotation
- **ktrdr/training/model_storage.py**: Improved Path Optional handling
- **ktrdr/neural/models/mlp.py**: Added Union type for optimizers
- **ktrdr/indicators/distance_from_ma_indicator.py**: Fixed MA indicator typing
- **ktrdr/config/models.py**: Fixed Pydantic Field default_factory patterns
- **ktrdr/ib/connection.py**: Added Future None checks for thread safety
- **ktrdr/api/endpoints/system.py**: Added explicit ApiResponse error parameter
- **ktrdr/training/model_trainer.py**: Fixed tensor operations type casting
- **ktrdr/api/services/gap_analysis_service.py**: Fixed model field usage
- **ktrdr/visualization/config_builder.py**: Fixed dynamic dict indexing

## Quality Assurance
- 1036 unit tests pass (1 unrelated visualization error)
- No functional regressions
- Maintained type safety improvements
- Systematic error categorization approach

## Running Total: 555 → 326 errors (229 fixed, 41.3% reduction)

+++ ktrdr/training/model_trainer.py
@@ -341,7 +341,7 @@ class ModelTrainer:
                     val_outputs = model(X_val)
                     val_loss = criterion(val_outputs, y_val)
                     _, val_predicted = torch.max(val_outputs.data, 1)
-                    val_accuracy = (val_predicted == y_val).float().mean().item()
+                    val_accuracy = float((val_predicted == y_val).float().mean().item())
                     val_loss = val_loss.item()
 
                 # Save best model
@@ -694,7 +694,7 @@ class ModelTrainer:
 
                     val_loss = criterion(val_outputs, y_val)
                     _, val_predicted = torch.max(val_outputs.data, 1)
-                    val_accuracy = (val_predicted == y_val).float().mean().item()
+                    val_accuracy = float((val_predicted == y_val).float().mean().item())
                     val_loss = val_loss.item()
 
                 # Save best model
